[EMMA-61] Production EMMA service - In Ingest::Message::Response and AwsS3::Message::Response, moved response   error processing to Api::Shared::ErrorTable. - Corrected #make_error_table to support all error messages, even ones not   directly mapped to a item. - In Workflow::Base::Events fixed mechanism for overrides of Workflow event   methods. - In Workflow::Base::Data renamed :failed to :failures. - Changed Workflow use of @succeeded and @failures to self.succeeded and   self.failures. - In UploadWorkflow::External#update_in_index amd #remove_from_index, moved   error message processing to #process_ingest_errors, which can make use of   general error messages (e.g. ones from Faraday) as well as error messages   which indicate problems with specific items.

+++ app/models/upload/file_methods.rb
@@ -223,6 +223,7 @@ module Upload::FileMethods
   #
   def delete_file(no_raise: true, field: nil)
     __debug_args(binding)
+    return if destroyed?
     if field.nil? && active_attached_file
       active_file_attacher.destroy
       active_file_attacher.set(nil)
